Moved to lucene package org.elasticsearch.lucene.spatial
