+++ server/src/test/java/org/elasticsearch/lucene/spatial/CartesianCentroidCalculatorTests.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.index.fielddata;
+package org.elasticsearch.lucene.spatial;
 
 import org.elasticsearch.geo.ShapeTestUtils;
 import org.elasticsearch.geometry.Line;
