+++ server/src/test/java/org/elasticsearch/lucene/spatial/CartesianShapeCoordinateEncoderTests.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.index.fielddata;
+package org.elasticsearch.lucene.spatial;
 
 import org.apache.lucene.geo.XYEncodingUtils;
 import org.elasticsearch.test.ESTestCase;
