+++ server/src/test/java/org/elasticsearch/lucene/spatial/GeoCentroidCalculatorTests.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.index.fielddata;
+package org.elasticsearch.lucene.spatial;
 
 import org.elasticsearch.geo.GeometryTestUtils;
 import org.elasticsearch.geometry.Line;
