+++ server/src/test/java/org/elasticsearch/lucene/spatial/GeoShapeCoordinateEncoderTests.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.index.fielddata;
+package org.elasticsearch.lucene.spatial;
 
 import org.apache.lucene.geo.GeoEncodingUtils;
 import org.elasticsearch.test.ESTestCase;
