+++ server/src/test/java/org/elasticsearch/lucene/spatial/TriangleTreeTests.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.index.fielddata;
+package org.elasticsearch.lucene.spatial;
 
 import org.apache.lucene.index.IndexableField;
 import org.apache.lucene.util.BytesRef;
