+++ x-pack/plugin/spatial/src/main/java/org/elasticsearch/xpack/spatial/index/fielddata/Component2DRelationVisitor.java
@@ -9,13 +9,13 @@ package org.elasticsearch.xpack.spatial.index.fielddata;
 
 import org.apache.lucene.geo.Component2D;
 import org.apache.lucene.index.PointValues;
-import org.elasticsearch.index.fielddata.CoordinateEncoder;
-import org.elasticsearch.index.fielddata.TriangleTreeWriter;
+import org.elasticsearch.lucene.spatial.CoordinateEncoder;
+import org.elasticsearch.lucene.spatial.TriangleTreeWriter;
 
-import static org.elasticsearch.index.fielddata.TriangleTreeVisitor.TriangleTreeDecodedVisitor;
-import static org.elasticsearch.index.fielddata.TriangleTreeVisitor.abFromTriangle;
-import static org.elasticsearch.index.fielddata.TriangleTreeVisitor.bcFromTriangle;
-import static org.elasticsearch.index.fielddata.TriangleTreeVisitor.caFromTriangle;
+import static org.elasticsearch.lucene.spatial.TriangleTreeVisitor.TriangleTreeDecodedVisitor;
+import static org.elasticsearch.lucene.spatial.TriangleTreeVisitor.abFromTriangle;
+import static org.elasticsearch.lucene.spatial.TriangleTreeVisitor.bcFromTriangle;
+import static org.elasticsearch.lucene.spatial.TriangleTreeVisitor.caFromTriangle;
 
 /**
  * A reusable tree reader visitor for a previous serialized {@link org.elasticsearch.geometry.Geometry} using
