+++ x-pack/plugin/spatial/src/test/java/org/elasticsearch/xpack/spatial/index/mapper/GeoShapeScriptMapperTests.java
@@ -14,11 +14,11 @@ import org.apache.lucene.index.IndexableField;
 import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.util.BytesRef;
 import org.elasticsearch.geometry.Point;
-import org.elasticsearch.index.fielddata.CentroidCalculator;
-import org.elasticsearch.index.fielddata.CoordinateEncoder;
-import org.elasticsearch.index.fielddata.GeometryDocValueWriter;
 import org.elasticsearch.index.mapper.MapperScriptTestCase;
 import org.elasticsearch.index.mapper.OnScriptError;
+import org.elasticsearch.lucene.spatial.CentroidCalculator;
+import org.elasticsearch.lucene.spatial.CoordinateEncoder;
+import org.elasticsearch.lucene.spatial.GeometryDocValueWriter;
 import org.elasticsearch.plugins.Plugin;
 import org.elasticsearch.script.GeometryFieldScript;
 import org.elasticsearch.search.lookup.SearchLookup;
